fix: Bleed applying through block. - Reduced network size for status effect updates. - Fixed bug where status effect's may not update gui on reset. - Removed entities applying multiple different status effects. - Reduced Toxin buildup from Swet latch. Burrukai's and Taegore's apply fracture. Varany's apply bleed. API: - Made it a little easier to apply status effect's. Now IAetherStatusEffect has a static method which checks capabilities and applies. - Updated all entities which utilize status effects to use this.

+++ src/main/java/com/gildedgames/aether/common/network/packets/effects/PacketStatusEffect.java
@@ -52,8 +52,8 @@ public class PacketStatusEffect implements IMessage
 		}
 		for (int i = 0; i < this.numberOfDirtyEffects; i++)
 		{
-			int effectId = buf.readShort();
-			int effectBuildup = buf.readShort();
+			int effectId = buf.readByte();
+			int effectBuildup = buf.readByte();
 			boolean isEffectApplied = buf.readBoolean();
 
 			this.statusEffectData.add(new StatusEffectData(effectId,effectBuildup,isEffectApplied));
@@ -68,8 +68,8 @@ public class PacketStatusEffect implements IMessage
 
 		for (StatusEffectData data : this.statusEffectData)
 		{
-			buf.writeShort(data.effectId);
-			buf.writeShort(data.buildup);
+			buf.writeByte(data.effectId);
+			buf.writeByte(data.buildup);
 			buf.writeBoolean(data.isApplied);
 		}
 	}
